chore(slug): update slug example stories to sb v7
